+++ src/main/java/com/example/demo/repository/ProfileRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProfileRepo extends ListCrudRepository<Profile,Integer> {
+public interface ProfileRepo extends ListCrudRepository<Profile,Long> {
 }
