+++ src/main/java/com/example/demo/repository/UserRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepo extends ListCrudRepository<User,Integer> {
+public interface UserRepo extends ListCrudRepository<User,Long> {
 }
