+++ src/main/java/com/example/demo/service/JobService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface JobService {
     JobDto save(JobDto JobDto);
     public List<JobDto> getAll();
-    JobDto getById(int id);
-    JobDto update(JobDto JobDto, int id);
-    boolean delete(int id);
+    JobDto getById(long id);
+    JobDto update(JobDto JobDto, long id);
+    boolean delete(long id);
 }
